chore: adding modules auth login & metadata
